Создаст дочерний презентер
